length for columns fixed

+++ src/main/kotlin/com/abdulkhalekomar/library_api/langauge/LanguageRepository.kt
@@ -2,4 +2,4 @@ package com.abdulkhalekomar.library_api.langauge
 
 import org.springframework.data.repository.CrudRepository
 
-interface LanguageRepository : CrudRepository<Language, Long>
+interface LanguageRepository : CrudRepository<Language, Int>
